001 not needed, do not suppress, e-items new location
The 001 field wasn't needed and was throwing an error when it was missing in the file.  References to the 001 removed.  Decided not to make ordered items discoverySuppress=true so they would be added to discovery during the next update.  Added a new property - permELocation - so electronic items could be assigned a different perm. location than physical items.