New tag, fixed validation method, cleanup
Fix to the validateRequiredValues method.  It was returning responseMessage prematurely on line 567 of the OrderImportShortened class.  It should only return at that point if an error has been found.  It should have been: if (!responseMessages.isEmpty()) return responseMessages;
Upgrade to java 11
New and optional tag was added.  It will look for it in the 980$r field.
General cleanup.